use size_t instead of uint64_t for rs_hash type

+++ include/libreset/hash.h
@@ -1,7 +1,7 @@
 #ifndef __LIBRESET_HASH_H__
 #define __LIBRESET_HASH_H__
 
-#include <stdint.h>
+#include <stddef.h>
 
 /**
  * The hash type
@@ -9,7 +9,7 @@
  * The `rs_hash` type is guaranteed to be able to be handled with the normal
  * comparators.
  */
-typedef rs_hash uint64_t;
+typedef rs_hash size_t;
 
 #endif //__LIBRESET_HASH_H__
